==> 01-scRNAseq Acute Wounds/05_subClustering/s3-Angiogenic/s2_clean_OtherCells_mg_GOs.Rmd <==
---
title: "Marker genes and DE analysis res0.8"
author: "Zhuang Liu"
date: "02/12/2021 Updated: `r Sys.Date()`"
output:
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width=12, fig.height = 8, warning=FALSE, message=FALSE)
```

```{r warning=FALSE, message=FALSE}
library(Seurat)
library(SeuratObject)
library(SeuratDisk)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(cowplot)
library(harmony)
```

####################################################
# Step. Marker genes identification
####################################################
## S. Known marker genes
```{r fig.width=9, fig.height=4}
hswound.endo.com <- readRDS("s1_clean_othercell.rds")
hswound.endo.com
DimPlot(hswound.endo.com, reduction = "umap", group.by='SCT_snn_res.0.8', label=TRUE, label.size = 6, split.by = "Condition") + NoAxes()
table(hswound.endo.com$SCT_snn_res.0.8, hswound.endo.com$Condition)
DimPlot(hswound.endo.com, reduction = "umap", group.by='SCT_snn_res.0.8', label=TRUE, label.size = 6) + FeaturePlot(hswound.endo.com, features = "MKI67", split.by = 'Condition', order = T, combine = F)

```

```{r fig.height=14}
plotcom <- FeaturePlot(
  hswound.endo.com,
  features=c('COL1A1', 'COL1A2', 'KRT5', 'KRT10', 'CD79A', 'CD7', 'CD3D', 
             "JCHAIN", "MZB1", 'CD4', 'CD8A', 'NKG7', 'HLA-DRA', 'CD68', 'CD163','PECAM1', "PMEL"), 
  combine=F, cols = c("grey","red"), order = T
)

# apply theme to each feature plot
for(i in 1:length(plotcom)){
  plotcom[[i]] <- plotcom[[i]] + NoLegend() + NoAxes()
}

cowplot::plot_grid(plotlist = plotcom, ncol = 4)
```

```{r}
DimPlot(object = hswound.endo.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "orig.ident", label.size = 4.5) + NoAxes() 

table(hswound.endo.com$SCT_snn_res.0.8)
DimPlot(object = hswound.endo.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.0.8", label.size = 4.5) + NoAxes() 

DimPlot(object = hswound.endo.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "newCellTypes", label.size = 5) + NoAxes() + ggtitle("")
```

```{r fig.height=4}
DimPlot(hswound.endo.com, reduction = "umap", pt.size = .001,group.by = "SCT_snn_res.0.8", split.by = 'Condition', ncol = 4, label = F) + NoLegend()+ ggtitle("")
DimPlot(hswound.endo.com, reduction = "umap", pt.size = .001,group.by = "newCellTypes", split.by = 'Condition', ncol = 4, label = F) + NoLegend()+ ggtitle("")
FeaturePlot(hswound.endo.com, features = "MKI67", split.by = 'Condition', order = T)
```

```{r fig.width=6}
DefaultAssay(hswound.endo.com) <- "SCT"
hswound.endo.com@active.ident <- hswound.endo.com$SCT_snn_res.0.8
hswound.endo.com <- BuildClusterTree(object = hswound.endo.com, assay = "SCT", slot = "scale.data")
suppressMessages(library(scales))
cluster.color <- hue_pal()(length(levels(hswound.endo.com)))
# pull the tree
hswound.clustertree <- Tool(object = hswound.endo.com, slot = "BuildClusterTree")
# plot the tree
suppressMessages(require(ape))
{plot.phylo(x = hswound.clustertree, direction = "rightwards", use.edge.length = TRUE, font = 4, tip.color = cluster.color, edge.width = 2, label.offset = 5, no.margin = TRUE)
tiplabels(pch = 19, col = cluster.color, adj = 3, cex = 2)}
```

## S. Visualizing marker genes for each cluster
```{r fig.width=12, fig.height=30}
DefaultAssay(hswound.endo.com) <- "RNA"
markers <- data.table::fread("s1_clean_othercell_mg_res0.8.txt")

markers_top10 <- markers %>% group_by(cluster) %>% top_n(20, avg_log2FC)
fac_levs <- unique(markers$cluster)

plot_list <- list()
for (i in seq_along(fac_levs)) {
  #print(paste0("Cluster ", i, " :Top 10 marker genes"))
  names <- fac_levs[i]
  plot_list[[i]] <- DotPlot(hswound.endo.com, features = markers_top10[markers_top10$cluster == names, ]$gene, col.max = 5, col.min = 0, 
                                dot.scale = 4, dot.min = 0.1,
                                assay = "RNA", group.by = "SCT_snn_res.0.8") + RotatedAxis() + coord_flip() + xlab("Marker genes") + ylab("") +
    theme(axis.text.x=element_text(size=10), axis.text.y=element_text(size=10)) + ggtitle(label = names)
}

wrap_plots(plot_list,ncol = 2) #+ plot_layout(guides = 'collect') & theme(legend.position = 'right')
```

## S. featurePlot Visualizing marker genes for each cluster
```{r fig.height=14}
markers_top10 <- markers %>% group_by(cluster) %>% top_n(30, avg_log2FC)
fac_levs <- unique(markers$cluster)

for (i in seq_along(fac_levs)) {
  print(paste0("Plot cluster", i-1))
  names <- fac_levs[i]
  plotgenes= markers_top10[markers_top10$cluster == names, ]$gene
  plot_list <- FeaturePlot(
    hswound.endo.com,
    features=plotgenes, order = T,
    combine=FALSE, cols = c("gray","red")
  )
  # apply theme to each feature plot
  for(i in 1:length(plot_list)){
    plot_list[[i]] <- plot_list[[i]] + NoLegend() + NoAxes()
  }
  print(wrap_plots(plot_list, ncol = 5))
}
```

```{r}
FeaturePlot(hswound.endo.com, features = c("VEGFA"))
```

# SessionInfo
```{r}
sessionInfo()
```
