==> 01-scRNAseq Acute Wounds/05_subClustering/s3-Angiogenic/s3_clean_OtherCells_updateAnnos.Rmd <==
---
title: "Sub-clustering of Fibroblasts"
author: "Zhuang Liu"
date: "30/11/2021 Updated: `r Sys.Date()`"
output:
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width = 12, warning = FALSE, message = FALSE)
```

```{r warning=FALSE, message=FALSE}
library(Seurat)
library(SeuratObject)
library(SeuratDisk)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(harmony)
```

####################################################
# Step. Plots
####################################################
## S. Load the data
```{r fig.width=8, fig.height=6}
hswound.endo.com <- readRDS("s1_clean_endothelial.rds") 
allEndoCells_clean_CellID <- colnames(hswound.endo.com)
saveRDS(allEndoCells_clean_CellID, file = "allEndoCells_clean_CellID.rds")

DimPlot(hswound.endo.com, reduction = "umap", group.by='SCT_snn_res.0.8', label=TRUE, label.size = 8) + NoAxes()

FeaturePlot(
  hswound.endo.com,
  features=c("CD4", "FCGR2A", "MRC1", "MS4A4A"),
  combine=T, cols = c("gray","red")
)
```

## FeaturePlot (Single)
```{r}
canonical_markers <- c("PECAM1", "EGFL7", "ENG", "IGFBP3", "FN1", "RBP7", "RGCC", "COL15A1", "ICAM1", "SELE", "MT2A", "CYP1B1", "SOX17", "PLAUR", "SNCG")

#rownames(hswound.endo.com)[grep("^HB", rownames(hswound.endo.com))]
plot_list <- FeaturePlot(
  hswound.endo.com,
  features=unlist(canonical_markers),
  combine=FALSE, cols = c("gray","red")
)

# apply theme to each feature plot
for(i in 1:length(plot_list)){
  plot_list[[i]] <- plot_list[[i]] + NoAxes()
}

wrap_plots(plot_list, ncol = 3)
```

## Dotplot
```{r eval=TRUE}
canonical_markers <- c("CCL21", "TFF3", "G0S2", "CSF3", "ID1", "SOCS3", "VCAM1", "IL6", "CYP1B1", "RGCC", "RFLNB", "IGFBP3", "CXCL12", "SEMA3G","RBP7")
fac_levs <- seq(0,7)
#hswound.endo.com$SCT_snn_res.0.8 <- factor(hswound.endo.com$SCT_snn_res.0.8, levels = rev(fac_levs))

DotPlot(hswound.endo.com, features = canonical_markers, 
        group.by = "SCT_snn_res.0.8", cols = c("lightgrey", "red"), 
        dot.scale = 4, col.min = 0, dot.min = 0.1
        ) + theme(axis.text.x = element_text(angle = 45, hjust = 1)) + labs(x="", y="")


plot_list <- FeaturePlot(
  hswound.endo.com,
  features=unlist(canonical_markers),
  combine=FALSE, cols = c("gray","red")
)

# apply theme to each feature plot
for(i in 1:length(plot_list)){
  plot_list[[i]] <- plot_list[[i]] + NoLegend() + NoAxes()
}

wrap_plots(plot_list, ncol = 8)
```

## cell proportion
```{r}
hswound.endo.com@active.ident <- hswound.endo.com$SCT_snn_res.0.8
clusters <- unique(hswound.endo.com@meta.data[["SCT_snn_res.0.8"]])
df <- data.frame()
for(i in 1:length(clusters)){
  SmCell_sum <- table(hswound.endo.com$orig.ident)
  tmp.df1 <- hswound.endo.com@meta.data %>% subset(SCT_snn_res.0.8 == clusters[i]) %>% select(orig.ident) %>% table()
  if(length(tmp.df1) == 12){
    #First normalize the total cell counts per sample
    cur_df <- as.data.frame(tmp.df1 / SmCell_sum)
    colnames(cur_df) <- c("Sample", "Freq")
    #Calculate the normalized proportion 
    cur_df$Freq <- cur_df$Freq * 1/(sum(cur_df$Freq))
    cur_df$Cluster <- clusters[i]
    df <- rbind(df, cur_df)
  } else {
    #print(i);print(tmp.df1)
    #only include the matched samples
    match.sample <- SmCell_sum[names(SmCell_sum) %in% names(tmp.df1)]
    #print(match.sample)
    #First normalize the total cell counts per sample
    cur_df <- as.data.frame(tmp.df1 / match.sample)
    colnames(cur_df) <- c("Sample", "Freq")
    #Calculate the normalized proportion 
    cur_df$Freq <- cur_df$Freq * 1/(sum(cur_df$Freq))
    cur_df$Cluster <- clusters[i]
    df <- rbind(df, cur_df)
  }
}

df.group <- df %>% mutate(Sample = gsub("PWH..", "", Sample)) %>% group_by(Cluster, Sample) %>% 
  summarise(Freq = sum(Freq)) %>% mutate(Group = Sample) %>% ungroup()
df.group$Group <- gsub("D0", "Skin", df.group$Group)
df.group$Group <- gsub("D1", "Wound1", df.group$Group)
df.group$Group <- gsub("D7", "Wound7", df.group$Group)
df.group$Group <- gsub("D30", "Wound30", df.group$Group)

df.group$Group <- factor(df.group$Group, levels = c("Skin", "Wound1", "Wound7", "Wound30"))
df.group$Cluster <- factor(df.group$Cluster, levels = fac_levs)
#df.group.f <- df.group[!df.group$Cluster %in% c(1,2,3,6,7),]
ggplot(df.group, aes(x = Cluster, y = Freq, fill = Group)) +
  geom_bar(position = "stack", stat = "identity") +
  scale_fill_brewer(palette = "Dark2") + # Paired or Set3 maximum variable = 12
  xlab('Cell Types') +
  scale_y_continuous(breaks = seq(0, 1, 0.1),
                     expand = c(0, 0),
                     name = 'Percentage') +
  geom_hline(yintercept=0.25, linetype="dashed", color = "black") +
  geom_hline(yintercept=0.5, linetype="dashed", color = "black") +
  geom_hline(yintercept=0.75, linetype="dashed", color = "black") +
  theme_bw() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    strip.background = element_blank(),
    #panel.border = element_rect(size = 0.35),
    axis.ticks = element_blank(),
    axis.text.x = element_text(size = 14, color = "black", angle = 45, vjust = 0.6),
    axis.text.y = element_text(size = 12, color = "black"),
    axis.title.y = element_text(size = 16, color = "black")
  ) 

```

## prepare file for scvelo
```{r}
hswound.endo.com <- readRDS("s1_clean_endothelial.rds") 
DefaultAssay(hswound.endo.com) <- "RNA"
# save metadata table: (first change the data type of cell clusters)
hswound.endo.com$nCount_SCT <- NULL
hswound.endo.com$nFeature_SCT <- NULL
hswound.endo.com$fil <- NULL
hswound.endo.com$orig.ident <- as.character(hswound.endo.com$orig.ident)
hswound.endo.com$Age <- as.character(hswound.endo.com$Age)
hswound.endo.com$Condition <- as.character(hswound.endo.com$Condition)
hswound.endo.com$SCT_snn_res.1.2 <- as.character(hswound.endo.com$SCT_snn_res.1.2)
hswound.endo.com$SCT_snn_res.0.5 <- as.character(hswound.endo.com$SCT_snn_res.0.5)
hswound.endo.com$SCT_snn_res.0.8 <- as.character(hswound.endo.com$SCT_snn_res.0.8)
hswound.endo.com$SCT_snn_res.1 <- as.character(hswound.endo.com$SCT_snn_res.1)
hswound.endo.com$seurat_clusters <- as.character(hswound.endo.com$seurat_clusters)
str(hswound.endo.com@meta.data)

hswound.endo.com$barcode <- as.character(colnames(hswound.endo.com))
hswound.endo.com$UMAP_1 <- hswound.endo.com@reductions$umap@cell.embeddings[,1]
hswound.endo.com$UMAP_2 <- hswound.endo.com@reductions$umap@cell.embeddings[,2]
write.csv(hswound.endo.com@meta.data, file='s4_scVelo_preparation/metadata.csv', quote=F, row.names=F)

# write expression counts matrix
library(Matrix)
counts_matrix <- GetAssayData(hswound.endo.com, assay='RNA', slot='counts')
writeMM(counts_matrix, file='s4_scVelo_preparation/counts.mtx')

# write dimesnionality reduction matrix, pca and harmony
write.csv(hswound.endo.com@reductions$pca@cell.embeddings, file='s4_scVelo_preparation/pca.csv', quote=F, row.names=F)

write.csv(hswound.endo.com@reductions$harmony@cell.embeddings, file='s4_scVelo_preparation/harmony.csv', quote=F, row.names=F)

# write gene names
write.table(
  data.frame('gene'=rownames(counts_matrix)),file='s4_scVelo_preparation/gene_names.csv',
  quote=F,row.names=F,col.names=F)
```

# SessionInfo
```{r}
sessionInfo()
```
